+++ zi/src/lib.rs
@@ -96,24 +96,31 @@
 //!         ShouldRender::Yes
 //!     }
 //!
-//!     // Whether the component is currently focused which will caused
-//!     // `input_binding` to be called on user input.
-//!     fn has_focus(&self) -> bool {
-//!         true
-//!     }
+//!    // Updates the key bindings of the component.
+//!    //
+//!    // This method will be called after the component lifecycle methods. It is
+//!    // used to specify how to react in response to keyboard events, typically
+//!    // by sending a message.
+//!    fn bindings(&self, bindings: &mut Bindings<Self>) {
+//!        // If we already initialised the bindings, nothing to do -- they never
+//!        // change in this example
+//!        if !bindings.is_empty() {
+//!            return;
+//!        }
+//!        // Set focus to `true` in order to react to key presses
+//!        bindings.set_focus(true);
 //!
-//!     // Send messages in reaction to user input.
-//!     fn input_binding(&self, pressed: &[Key]) -> BindingMatch<Self::Message> {
-//!         BindingMatch::clear(match pressed {
-//!             &[Key::Char('+')]  => Some(Message::Increment),
-//!             &[Key::Char('-')] => Some(Message::Decrement),
-//!             &[Key::Esc] => {
-//!                 self.link.exit();
-//!                 None
-//!             }
-//!             _ => None,
-//!         })
-//!     }
+//!        // Increment
+//!        bindings.add("increment", [Key::Char('+')], || Message::Increment);
+//!        bindings.add("increment", [Key::Char('=')], || Message::Increment);
+//!
+//!        // Decrement
+//!        bindings.add("decrement", [Key::Char('-')], || Message::Decrement);
+//!
+//!        // Exit
+//!        bindings.add("exit", [Key::Ctrl('c')], |this: &Self| this.link.exit());
+//!        bindings.add("exit", [Key::Esc], |this: &Self| this.link.exit());
+//!    }
 //! }
 //!
 //! fn main() -> zi_term::Result<()> {
@@ -129,18 +136,19 @@ pub mod components;
 pub mod terminal;
 
 pub use component::{
+    bindings::{AnyChar, BindingQuery, Bindings, Keymap},
     layout::{self, ComponentExt, ComponentKey, Container, FlexBasis, FlexDirection, Item},
-    BindingMatch, BindingTransition, Callback, Component, ComponentLink, Layout, ShouldRender,
+    Callback, Component, ComponentLink, Layout, ShouldRender,
 };
 pub use terminal::{Background, Canvas, Colour, Foreground, Key, Position, Rect, Size, Style};
 
 pub mod prelude {
     //! The Zi prelude.
-    pub use super::{Background, Canvas, Colour, Foreground, Key, Position, Rect, Size, Style};
     pub use super::{
-        BindingMatch, BindingTransition, Component, ComponentExt, ComponentLink, Container,
-        FlexBasis, FlexDirection, Item, Layout, ShouldRender,
+        AnyChar, Bindings, Component, ComponentExt, ComponentLink, Container, FlexBasis,
+        FlexDirection, Item, Layout, ShouldRender,
     };
+    pub use super::{Background, Canvas, Colour, Foreground, Key, Position, Rect, Size, Style};
 }
 
 // Crate only modules